third_party/blink/web_tests/fast/events/menu-key-context-menu-document-expected.txt

Target

This test checks if contextmenu event target is correct when handled at the document level. To test manually, first right-click on the "Target" text in the orange box, then press the menu key. Both interactions should result in the message: "PASS anchor is anchor".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS anchor is anchor
PASS anchor is anchor
PASS successfullyParsed is true

TEST COMPLETE
